Fixed ory whoami request every time on sidebar-click
